feat: repository simplistic example
