+++ lib/features/todo/todo_page.dart
@@ -1,0 +1,134 @@
+import 'package:flutter/material.dart';
+import 'package:todo/features/todo/todo_page_view_model.dart';
+import 'package:todo/main.dart';
+import 'package:todo/shared/ui_utilities/value_listenable_builder_x.dart';
+import 'package:todo/todo.dart';
+
+class TodoPage extends StatefulWidget {
+  const TodoPage({super.key});
+
+  @override
+  State<TodoPage> createState() => _TodoPageState();
+}
+
+class _TodoPageState extends State<TodoPage> {
+  late final homePageViewModel = TodoPageViewModel(
+    dateService: dateService,
+    todoRepository: todoRepository,
+  );
+
+  final TextEditingController _titleController = TextEditingController();
+  final TextEditingController _descriptionController = TextEditingController();
+
+  @override
+  void initState() {
+    super.initState();
+    homePageViewModel.init();
+  }
+
+  @override
+  void dispose() {
+    _titleController.dispose();
+    _descriptionController.dispose();
+    homePageViewModel.dispose();
+    super.dispose();
+  }
+
+  void _addTodo() {
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text("Add Todo"),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              TextField(
+                controller: _titleController,
+                decoration: const InputDecoration(
+                  labelText: "Title",
+                ),
+              ),
+              TextField(
+                controller: _descriptionController,
+                decoration: const InputDecoration(
+                  labelText: "Description",
+                ),
+              ),
+            ],
+          ),
+          actions: [
+            TextButton(
+              onPressed: () {
+                homePageViewModel.add(
+                  Todo(
+                    title: _titleController.text,
+                    description: _descriptionController.text,
+                  ),
+                );
+
+                _titleController.clear();
+                _descriptionController.clear();
+                Navigator.pop(context);
+              },
+              child: const Text("Add"),
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: ValueListenableBuilder(
+          valueListenable: homePageViewModel.serviceDate,
+          builder: (context, date, child) {
+            return Text("Todo $date");
+          },
+        ),
+        actions: [
+          ValueListenableBuilder2(
+            first: homePageViewModel.todos,
+            second: homePageViewModel.showCompletedTodos,
+            builder: (context, todos, showCompletedTodos, child) {
+              if (homePageViewModel.hasNonCompletedTodos) {
+                return TextButton(
+                  onPressed: () {
+                    homePageViewModel.toggleCompletedTodos();
+                  },
+                  child: showCompletedTodos
+                      ? const Text("Hide Done")
+                      : const Text("Show Done"),
+                );
+              }
+              return const SizedBox();
+            },
+          ),
+        ],
+      ),
+      body: TodoList(
+        toggleDone: homePageViewModel.toggleDone,
+        removeTodo: homePageViewModel.remove,
+        todosNotifier: homePageViewModel.todos,
+        showCompletedTodos: homePageViewModel.showCompletedTodos,
+      ),
+      floatingActionButton: Row(
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: [
+          FloatingActionButton(
+            onPressed: homePageViewModel.updateServiceDate,
+            child: const Icon(Icons.date_range),
+          ),
+          const SizedBox(width: 12),
+          FloatingActionButton(
+            onPressed: _addTodo,
+            child: const Icon(Icons.add),
+          ),
+        ],
+      ),
+    );
+  }
+}
