+++ lib/shared/date_service.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/foundation.dart';
+
+class DateService {
+  ValueNotifier<DateTime> dateNotifier = ValueNotifier(DateTime.now());
+
+  void updateDate() {
+    dateNotifier.value = DateTime.now();
+  }
+}
