+++ lib/shared/ui_utilities/value_listenable_builder_x.dart
@@ -1,0 +1,70 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+
+class ValueListenableBuilder2<A, B> extends StatelessWidget {
+  const ValueListenableBuilder2({
+    required this.first,
+    required this.second,
+    super.key,
+    required this.builder,
+    this.child,
+  });
+
+  final ValueListenable<A> first;
+  final ValueListenable<B> second;
+  final Widget? child;
+  final Widget Function(BuildContext context, A a, B b, Widget? child) builder;
+
+  @override
+  Widget build(BuildContext context) {
+    return ValueListenableBuilder<A>(
+      valueListenable: first,
+      builder: (_, a, __) {
+        return ValueListenableBuilder<B>(
+          valueListenable: second,
+          builder: (context, b, __) {
+            return builder(context, a, b, child);
+          },
+        );
+      },
+    );
+  }
+}
+
+class ValueListenableBuilder3<A, B, C> extends StatelessWidget {
+  const ValueListenableBuilder3({
+    required this.first,
+    required this.second,
+    required this.third,
+    super.key,
+    required this.builder,
+    this.child,
+  });
+
+  final ValueListenable<A> first;
+  final ValueListenable<B> second;
+  final ValueListenable<C> third;
+  final Widget? child;
+  final Widget Function(BuildContext context, A a, B b, C c, Widget? child)
+      builder;
+
+  @override
+  Widget build(BuildContext context) {
+    return ValueListenableBuilder<A>(
+      valueListenable: first,
+      builder: (_, a, __) {
+        return ValueListenableBuilder<B>(
+          valueListenable: second,
+          builder: (context, b, __) {
+            return ValueListenableBuilder<C>(
+              valueListenable: third,
+              builder: (context, c, __) {
+                return builder(context, a, b, c, child);
+              },
+            );
+          },
+        );
+      },
+    );
+  }
+}
